% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sentiment.R
\name{as_article_sentiment}
\alias{as_article_sentiment}
\title{Perform sentiment analysis on scraped articles}
\usage{
as_article_sentiment(article_text)
}
\arguments{
\item{article_text}{The text of the articles to analyze}
}
\value{
A data frame of sentiment scores
}
\description{
Perform sentiment analysis on scraped articles
}
