% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schemas.R
\name{check_schema}
\alias{check_schema}
\title{Check if a schema already exists for a website}
\usage{
check_schema(website_url)
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}
}
\value{
TRUE if the schema exists, FALSE otherwise.
}
\description{
This function checks if a schema CSV exists locally for the given website URL.
}
