% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_store_base.R
\name{compose_storage}
\alias{compose_storage}
\title{Compose a storage pipeline from multiple decorators}
\usage{
compose_storage(...)
}
\arguments{
\item{...}{Decorator functions (e.g. store_index_json, store_monthly_json)}
}
\value{
A composed function taking (data, news_site, overwrite)
}
\description{
STACK EM LIKE LEGOS YO
}
