% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_schema_elements.R
\name{find_schema_manual}
\alias{find_schema_manual}
\title{Manually find and add schema elements for a website}
\usage{
find_schema_manual(website_url)
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}
}
\value{
A list containing the schema elements (author, title, published date, article text).
}
\description{
This function allows the user to manually input the schema elements (author, title, published date, article text) for a given website.
}
