% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schemas.R
\name{gs_find_article_elements}
\alias{gs_find_article_elements}
\title{Automatically find and add schema elements for a website}
\usage{
gs_find_article_elements(website_url)
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}
}
\value{
A list containing the schema elements (author, title, published date, article text).
}
\description{
This function attempts to automatically find schema elements (author, title, published date, article text) for the given website.
}
