% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schemas.R
\name{gs_get_sitemap_str}
\alias{gs_get_sitemap_str}
\title{title}
\usage{
gs_get_sitemap_str(sitemap_url)
}
\arguments{
\item{sitemap_url}{parameter desc}
}
\value{
what it returns
}
\description{
description
}
