% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schemas.R
\name{gs_pull_schema}
\alias{gs_pull_schema}
\title{Pull schema elements for a website}
\usage{
gs_pull_schema(website_url)
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}
}
\value{
A list of schema elements (author_element, title_element, date_element, text_element) if available, NULL otherwise.
}
\description{
This function retrieves the schema elements for a given website from the local CSV file.
}
