% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schemas.R
\name{gs_remove_schema}
\alias{gs_remove_schema}
\title{Remove Schema Row(s) from the Local CSV File}
\usage{
gs_remove_schema(website_url, id = NULL, every = FALSE)
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}

\item{id}{An optional character string specifying the unique ID of the schema to remove.}

\item{every}{Logical; if TRUE, deletes all rows returned by \code{gs_pull_schema}. Default is FALSE.}
}
\value{
TRUE if the row(s) are successfully removed and the updated CSV file is written, FALSE otherwise.
}
\description{
This function removes schema row(s) from the local CSV file based on the specified \code{website_url} and optional \code{id}.
If \code{every = TRUE}, it deletes all rows returned by \code{gs_pull_schema}.
If \code{every = FALSE} (default), it prints an error if multiple rows are found and returns without deleting.
}
