% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schemas.R
\name{gs_rm_schema}
\alias{gs_rm_schema}
\title{title}
\usage{
gs_rm_schema(website_url)
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}
}
\value{
what it returns
}
\description{
description
}
