% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schemas.R
\name{gs_write_schema}
\alias{gs_write_schema}
\title{Write a New Schema to the Local CSV File}
\usage{
gs_write_schema(
  website_structure,
  starting_sitemap,
  author_element,
  title_element,
  date_element,
  text_element,
  structure,
  layer1_type = NA,
  layer1_class = NA,
  layer2_type = NA,
  layer2_class = NA,
  layer3_type = NA,
  layer3_class = NA,
  layer4_type = NA,
  layer4_class = NA
)
}
\arguments{
\item{website_structure}{A character string for the URL of the website.}

\item{starting_sitemap}{A character string for the URL of the starting sitemap.}

\item{author_element}{A character string for the CSS selector/XPath for the author element.}

\item{title_element}{A character string for the CSS selector/XPath for the title element.}

\item{date_element}{A character string for the CSS selector/XPath for the published date element.}

\item{text_element}{A character string for the CSS selector/XPath for the article text element.}

\item{structure}{A character string for the sitemap structure (e.g., nested, flat).}

\item{layer1_type, }{layer1_class, layer2_type, layer2_class, layer3_type, layer3_class, layer4_type, layer4_class Element types and classes for sitemap layers.}
}
\value{
TRUE if the schema is successfully written, FALSE otherwise.
}
\description{
This function writes a new schema to the local CSV file.
}
