% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_append_links}
\alias{gu_append_links}
\title{Append Sitemap URL to Year Links}
\usage{
gu_append_links(url_prefix, link_list)
}
\arguments{
\item{url_prefix}{A character string representing the base URL of the sitemap (e.g., "https://www.example.com/sitemap/").}

\item{link_list}{A character vector containing relative year links (e.g., "1990/", "1991/").}
}
\value{
A character vector of full URLs.
}
\description{
This function takes the base sitemap URL and appends it to each year link to create full URLs.
}
