% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_article_links}
\alias{gu_article_links}
\title{Get Article Links}
\usage{
gu_article_links(day_link, tag_type = NULL, tag_class = NULL)
}
\arguments{
\item{day_link}{A character string representing the URL of a link containing article links.}

\item{tag_type}{An optional character string specifying the type of HTML tag (e.g., "ol", "ul", "div"). Defaults to "ol".}

\item{tag_class}{An optional character string specifying the class of the tag containing links.}
}
\value{
A list of URLs of the article links.
}
\description{
Extracts a list of article links from an XML or HTML sitemap of a day.
}
