% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_day_links}
\alias{gu_day_links}
\title{title}
\usage{
gu_day_links(day_link, day_min, day_max)
}
\arguments{
\item{day_link}{A character string representing the URL of a link contaning day links}

\item{day_min}{Minimum day to include (numeric).}

\item{day_max}{Minimum day to include (numeric).}
}
\value{
A list of URLs of the day links
}
\description{
description
}
