% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_extract_sitemap_links}
\alias{gu_extract_sitemap_links}
\title{Extract Links from a Sitemap (Supports XML and HTML)}
\usage{
gu_extract_sitemap_links(sitemap_url)
}
\arguments{
\item{sitemap_url}{The URL of the sitemap.}
}
\value{
A list containing extracted URLs and a string indicating format type ("html" or "xml").
}
\description{
Fetches a sitemap, detects XML or HTML format, and extracts links.
}
