% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_get_article_links2}
\alias{gu_get_article_links2}
\title{Get Article Links from a Website}
\usage{
gu_get_article_links2(website_url, year_min, year_max, id = NULL)
}
\arguments{
\item{website_url}{A character string representing the URL of the website (e.g., "https://www.nytimes.com").}

\item{year_min}{The starting year of the range (e.g., 1990).}

\item{year_max}{The ending year of the range (e.g., 1991).}

\item{id}{A character string specifying the unique ID of the schema to pull (optional).}
}
\value{
A character vector of article links.
}
\description{
This function retrieves the list of article links from a given website and year range using schema details.
If multiple schema rows are returned from \code{gs_pull_schema}, it issues a warning to specify the \code{id} parameter.
}
