% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_get_layer_instructions}
\alias{gu_get_layer_instructions}
\title{Get Instructions for Each Layer of Sitemap Parsing}
\usage{
gu_get_layer_instructions(layer_info)
}
\arguments{
\item{layer_info}{A single-row data frame with columns "number", "type", and "class" for a layer.}
}
\value{
A character string with instructions, e.g., "single gu_parse_xml_sitemap_date_in_url" or "map gu_parse_xml_sitemap_date_in_tag".
}
\description{
This function determines the action and helper function for each layer
based on the schema info and whether the layer is the first or subsequent.
}
