% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_get_links}
\alias{gu_get_links}
\title{Get Links from a Website}
\usage{
gu_get_links(website_url, start_date, end_date, id = NULL)
}
\arguments{
\item{website_url}{A character string representing the URL of the website (e.g., "https://www.nytimes.com").}

\item{start_date}{The start date for retrieving article links (e.g., "2003-03-11").}

\item{end_date}{The end date for retrieving article links (e.g., "2004-12-31").}

\item{id}{A character string specifying the unique ID of the schema to pull (optional).}
}
\value{
A character vector of article links.
}
\description{
This function retrieves the list of article links from a given website and date range using schema details.
If multiple schema rows are returned from \code{gs_pull_schema}, it issues a warning to specify the \code{id} parameter.
}
