% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_parse_xml_sitemap_date_in_url}
\alias{gu_parse_xml_sitemap_date_in_url}
\title{Scrape Sitemap with Date in URL}
\usage{
gu_parse_xml_sitemap_date_in_url(sitemap_url, start_date, end_date)
}
\arguments{
\item{sitemap_url}{The URL of the XML sitemap to parse.}

\item{start_date}{The start date for filtering (as "YYYY-MM-DD").}

\item{end_date}{The end date for filtering (as "YYYY-MM-DD").}
}
\value{
A character vector of filtered links from the sitemap.
}
\description{
Parses an XML sitemap where the date information is part of the URL and filters links by a date range.
}
