% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_ymdl_str_links}
\alias{gu_ymdl_str_links}
\title{Scrape for links by date in a year-month-day-links website structure}
\usage{
gu_ymdl_str_links(sitemap_url, date_range)
}
\arguments{
\item{sitemap_url}{A character string representing the sitemap URL of the website.}

\item{date_range}{Range of dates to return urls for}
}
\value{
List of links for the specified date range
}
\description{
Use for websites with a year-month-day-links sitemap structure
}
