% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_schema_elements.R
\name{prompt_overwrite}
\alias{prompt_overwrite}
\title{Prompt the user to overwrite an existing schema}
\usage{
prompt_overwrite(website_url)
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}
}
\value{
TRUE if the user chooses to overwrite, FALSE otherwise.
}
\description{
This function prompts the user for permission to overwrite an existing schema in the local CSV file.
}
