% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sentiment_analysis.R
\name{run_sentiment_analysis}
\alias{run_sentiment_analysis}
\title{Perform sentiment analysis on scraped articles}
\usage{
run_sentiment_analysis(article_text)
}
\arguments{
\item{article_text}{The text of the articles to analyze}
}
\value{
A data frame of sentiment scores
}
\description{
Perform sentiment analysis on scraped articles
}
