% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_extract_author}
\alias{sa_extract_author}
\title{Extract Article Author}
\usage{
sa_extract_author(article_html)
}
\arguments{
\item{article_html}{An rvest HTML document object containing the full article content.}
}
\value{
A character string representing the author's name.
}
\description{
This function extracts the author's name from a scraped article HTML document.
}
