% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_extract_date}
\alias{sa_extract_date}
\title{Extract Published Date}
\usage{
sa_extract_date(article_html)
}
\arguments{
\item{article_html}{An rvest HTML document object containing the full article content.}
}
\value{
A character string representing the published date.
}
\description{
This function extracts the published date from a scraped article HTML document.
}
