% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_extract_text}
\alias{sa_extract_text}
\title{Extract Article Text}
\usage{
sa_extract_text(article_html, selector)
}
\arguments{
\item{article_html}{An rvest HTML document object.}

\item{selector}{The CSS/XPath selector for extracting the article text.}
}
\value{
A character string representing the cleaned article text or NA if not found.
}
\description{
Extract Article Text
}
