% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_extract_title}
\alias{sa_extract_title}
\title{Extract Article Title}
\usage{
sa_extract_title(article_html)
}
\arguments{
\item{article_html}{An rvest HTML document object containing the full article content.}
}
\value{
A character string representing the article title.
}
\description{
This function extracts the title from a scraped article HTML document.
}
