% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_get_article_data}
\alias{sa_get_article_data}
\title{Scrape and Extract Full Article Data}
\usage{
sa_get_article_data(article_url)
}
\arguments{
\item{article_url}{A character string representing the URL of the article to scrape.}
}
\value{
A data frame containing the article URL, title, author, published date, and text.
}
\description{
This function scrapes an article and extracts its key components, then organizes them into a dataframe.
}
