% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_get_selectors}
\alias{sa_get_selectors}
\title{Load Selectors for a Given News Website}
\usage{
sa_get_selectors(url)
}
\arguments{
\item{url}{A character string representing the article URL.}
}
\value{
A named list containing the CSS/XPath selectors for title, author, date, and text, or NULL if no custom selectors exist.
}
\description{
Load Selectors for a Given News Website
}
