% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_scrape_articles}
\alias{sa_scrape_articles}
\title{Scrape Multiple Articles}
\usage{
sa_scrape_articles(article_urls)
}
\arguments{
\item{article_urls}{A character vector containing multiple article URLs.}
}
\value{
A data frame where each row represents an article with its extracted data.
}
\description{
This function scrapes multiple articles from a list of URLs and returns a structured data frame.
}
