% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_new_articles.R
\name{scrape_new_articles}
\alias{scrape_new_articles}
\title{Scrape new articles from URL list}
\usage{
scrape_new_articles(website)
}
\arguments{
\item{website}{The website to scrape articles from}
}
\value{
A data frame of newly scraped articles
}
\description{
Scrape new articles from URL list
}
