% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{scrape_sitemap}
\alias{scrape_sitemap}
\title{scrapes sitemap for list of links in date range}
\usage{
scrape_sitemap(date_range)
}
\value{
list of links filtered by date range
}
\description{
scrapes sitemap for list of links in date range
}
