% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_website.R
\name{scrape_website}
\alias{scrape_website}
\title{Scrape website articles based on schema}
\usage{
scrape_website(website, url_list)
}
\arguments{
\item{website}{The website to scrape}

\item{url_list}{The list of URLs to scrape from}
}
\value{
A data frame of scraped articles
}
\description{
Scrape website articles based on schema
}
