% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_data.R
\name{sd_clean_date}
\alias{sd_clean_date}
\title{Clean Published Dates in Dataframe or List}
\usage{
sd_clean_date(data)
}
\arguments{
\item{data}{A dataframe or a list with at least a 'published_date' field.}
}
\value{
The same structure with 'published_date' cleaned.
}
\description{
Detects and standardizes the format of each date to "YYYY-MM-DD".
}
