% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_scraped_data.R
\name{ss_clean_date_text}
\alias{ss_clean_date_text}
\title{Clean Published Dates by Removing Unwanted Words}
\usage{
ss_clean_date_text(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing a "published_date" column.}
}
\value{
The same dataframe with a cleaned "published_date" column.
}
\description{
Removes all words that are not month names and removes commas.
}
