% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_store_decorators.R
\name{store_ensure_folders}
\alias{store_ensure_folders}
\title{make sure that the news site folder exists or else before writing files}
\usage{
store_ensure_folders(next_fn)
}
\description{
make sure that the news site folder exists or else before writing files
}
