% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_url_list.R
\name{store_url_list}
\alias{store_url_list}
\title{Store URL list for a website}
\usage{
store_url_list(website, url_list)
}
\arguments{
\item{website}{The website to scrape URLs from}

\item{url_list}{A list of URLs to store}
}
\value{
None
}
\description{
Store URL list for a website
}
