% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_url_list.R
\name{su_check_csv}
\alias{su_check_csv}
\title{title}
\usage{
su_check_csv(param1, param2)
}
\arguments{
\item{param1}{parameter desc}

\item{param2}{parameter desc}
}
\value{
what it returns
}
\description{
description
}
