% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_url_list.R
\name{su_pull_urls}
\alias{su_pull_urls}
\title{Pull URLs from CSV for a Specific News Website}
\usage{
su_pull_urls(start_date, end_date, website_url)
}
\arguments{
\item{start_date}{parameter desc}

\item{end_date}{param desc}

\item{website_url}{param desc}
}
\value{
A dataframe containing the title, published date, author, and text for a selected date range for a specific news website.
}
\description{
desc, too lazy do later
}
