% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_url_list.R
\name{su_read_csv}
\alias{su_read_csv}
\title{Read CSV and Filter Rows with NA Metadata}
\usage{
su_read_csv(website_url, folder_path = "inst/extdata/scraped_data/")
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}

\item{folder_path}{A character string specifying the folder where the CSV is located.}
}
\value{
A data frame containing the filtered rows, or an empty data frame if no matching rows are found.
}
\description{
This function reads a CSV file for a specific website and returns all rows
where the columns \code{published_date}, \code{author}, \code{title}, and \code{text} are all \code{NA}.
}
