% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_schema_elements.R
\name{write_schema}
\alias{write_schema}
\title{Write a new schema to the local CSV file}
\usage{
write_schema(website_url, schema_data)
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}

\item{schema_data}{A list containing the schema elements: author, title, published date, and article text.}
}
\value{
TRUE if the schema is successfully written, FALSE otherwise.
}
\description{
This function writes a new schema (author, title, published date, article text)
to the local CSV file if it does not already exist.
}
