% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sentiment.R
\name{as_article_sentiment}
\alias{as_article_sentiment}
\title{Perform Sentiment Analysis on Articles (Optimized & Fixed)}
\usage{
as_article_sentiment(dataframe)
}
\arguments{
\item{dataframe}{A data frame with a "text" column containing article text.}
}
\value{
The same data frame with additional columns: "sentiment_val" and "sentiment_sd".
}
\description{
Perform Sentiment Analysis on Articles (Optimized & Fixed)
}
