% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schemas.R
\name{gs_write_schema}
\alias{gs_write_schema}
\title{Write a new schema to the local CSV file}
\usage{
gs_write_schema(
  website_url,
  author_element,
  title_element,
  date_element,
  text_element,
  xml_structure
)
}
\arguments{
\item{website_url}{A character string for the URL of the website.}

\item{author_element}{A character string for the CSS selector/XPath for the author element.}

\item{title_element}{A character string for the CSS selector/XPath for the title element.}

\item{date_element}{A character string for the CSS selector/XPath for the published date element.}

\item{text_element}{A character string for the CSS selector/XPath for the article text element.}

\item{xml_structure}{A character string for the website structure}
}
\value{
TRUE if the schema is successfully written, FALSE otherwise.
}
\description{
This function writes a new schema to the local CSV file. If the schema already exists,
it prompts the user for confirmation to overwrite.
}
