% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_apply_day_links}
\alias{gu_apply_day_links}
\title{Get All Day Links from a List of Month Links}
\usage{
gu_apply_day_links(month_links, tag_type, tag_class)
}
\arguments{
\item{month_links}{A character vector containing the list of month URLs.}

\item{tag_type}{An optional character string specifying the type of HTML tag (e.g., "ol", "ul", "div"). Defaults to "ol".}

\item{tag_class}{An optional character string specifying the class of the tag containing links.}
}
\value{
A list of URLs of all the day links.
}
\description{
This function calls \code{gu_day_links} for each link in a list of month URLs and returns all day links.
}
