% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_day_links}
\alias{gu_day_links}
\title{Get Day Links}
\usage{
gu_day_links(day_link, tag_type = NULL, tag_class = NULL)
}
\arguments{
\item{day_link}{A character string representing the URL of a link containing day links.}

\item{tag_type}{An optional character string specifying the type of HTML tag (e.g., "ol", "ul", "div"). Defaults to "ol".}

\item{tag_class}{An optional character string specifying the class of the tag containing links.}
}
\value{
A list of URLs of the day links.
}
\description{
Extracts a list of day links from an XML or HTML sitemap.
}
