% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_day_links}
\alias{gu_day_links}
\title{Get Day Links}
\usage{
gu_day_links(
  month_link,
  day_min = 1,
  day_max = 31,
  tag_type = NULL,
  tag_class = NULL
)
}
\arguments{
\item{month_link}{A character string representing the URL of a month link containing day links.}

\item{day_min}{The starting day of the range (e.g., 1 for the 1st).}

\item{day_max}{The ending day of the range (e.g., 31 for the 31st).}

\item{tag_type}{An optional character string specifying the type of HTML tag (e.g., "ol", "ul", "div"). Defaults to "ol".}

\item{tag_class}{An optional character string specifying the class of the tag containing links.}
}
\value{
A list of URLs of the day links.
}
\description{
Extracts a list of day links within a specified range from an XML or HTML sitemap.
}
