% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_month_links}
\alias{gu_month_links}
\title{Get Month Links}
\usage{
gu_month_links(
  year_link,
  month_min = 1,
  month_max = 12,
  tag_type = NULL,
  tag_class = NULL
)
}
\arguments{
\item{year_link}{A character string representing the URL of a year link containing month links.}

\item{month_min}{The starting month of the range (e.g., 1 for January).}

\item{month_max}{The ending month of the range (e.g., 12 for December).}

\item{tag_type}{An optional character string specifying the type of HTML tag (e.g., "ol", "ul", "div"). Defaults to "ol".}

\item{tag_class}{An optional character string specifying the class of the tag containing links.}
}
\value{
A list of URLs of the month links.
}
\description{
Extracts a list of month links within a specified range from an XML or HTML sitemap.
}
