% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_list.R
\name{gu_month_links}
\alias{gu_month_links}
\title{Get Month Links}
\usage{
gu_month_links(month_link, month_min, month_max, ol_class = NULL)
}
\arguments{
\item{month_link}{A character string representing the URL of a link containing month links.}

\item{month_min}{Minimum month to include (numeric, 1-12).}

\item{month_max}{Maximum month to include (numeric, 1-12).}

\item{ol_class}{An optional character string specifying the class of the \if{html}{\out{<ol>}} tag containing links.}
}
\value{
A list of URLs of the month links.
}
\description{
Extracts a list of month links within a specified range from an XML or HTML sitemap.
}
