% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_extract_author}
\alias{sa_extract_author}
\title{Extract Article Author}
\usage{
sa_extract_author(article_html, selector)
}
\arguments{
\item{article_html}{An rvest HTML document object.}

\item{selector}{The CSS/XPath selector for extracting the author.}
}
\value{
A character string representing the author's name or NA if not found.
}
\description{
Extract Article Author
}
