% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_extract_title}
\alias{sa_extract_title}
\title{Extract Article Title}
\usage{
sa_extract_title(article_html, selector, tag)
}
\arguments{
\item{article_html}{An rvest HTML document object.}

\item{selector}{The CSS/XPath selector for extracting the title.}

\item{tag}{The CSS tag}
}
\value{
A character string representing the article title or NA if not found.
}
\description{
Extract Article Title
}
