% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_get_article_data}
\alias{sa_get_article_data}
\title{Scrape and Extract Full Article Data}
\usage{
sa_get_article_data(article_url, selectors = NULL)
}
\arguments{
\item{article_url}{A character string representing the URL of the article.}

\item{selectors}{A named list of CSS/XPath selectors and tags (can be NULL).}
}
\value{
A data frame containing the article data.
}
\description{
Scrape and Extract Full Article Data
}
