% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_articles.r
\name{sa_scrape_articles}
\alias{sa_scrape_articles}
\title{Scrape Multiple Articles}
\usage{
sa_scrape_articles(article_urls, verbose = TRUE)
}
\arguments{
\item{article_urls}{A character vector containing multiple article URLs.}

\item{verbose}{Logical; if TRUE, prints progress updates and execution time (default: TRUE).}
}
\value{
A data frame where each row represents an article.
}
\description{
Scrape Multiple Articles
}
