% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_data.R
\name{sd_clean_date}
\alias{sd_clean_date}
\title{Clean Published Dates in Dataframe}
\usage{
sd_clean_date(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing a "published_date" column.}
}
\value{
The same dataframe with "published_date" cleaned to "YYYY-MM-DD".
}
\description{
Detects the format of each date and converts it to "YYYY-MM-DD".
}
