% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_scraped_data.R
\name{ss_clean_date}
\alias{ss_clean_date}
\title{Clean Published Dates in Dataframe}
\usage{
ss_clean_date(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing a "published_date" column.}
}
\value{
The same dataframe with a cleaned "published_date" column.
}
\description{
Removes time and timezone, then converts dates into "YYYY-MM-DD" format.
}
