% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_scraped_data.R
\name{ss_pull_articles}
\alias{ss_pull_articles}
\title{Pull Articles from a News Site}
\usage{
ss_pull_articles(start_date, end_date, ran_articles = NULL, news_site)
}
\arguments{
\item{start_date}{The start date in YYYY-MM-DD format.}

\item{end_date}{The end date in YYYY-MM-DD format.}

\item{ran_articles}{The number of random articles to sample. If NULL, returns all articles (default: NULL).}

\item{news_site}{The name of the news site (e.g., "huffpost").}
}
\value{
A data frame containing the selected articles or an error message if conditions are not met.
}
\description{
Retrieves articles from a CSV for a given date range.
}
