% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_scraped_data.R
\name{ss_store_articles}
\alias{ss_store_articles}
\title{Store and Append Articles for Any News Site}
\usage{
ss_store_articles(
  article_data,
  news_site,
  folder_path = "inst/extdata/article_data/",
  overwrite = FALSE
)
}
\arguments{
\item{article_data}{A data frame with columns: url, title, author, published_date, text.}

\item{news_site}{The name of the news site (e.g., "huffpost").}

\item{folder_path}{Directory for storing CSV files.}

\item{overwrite}{TRUE or FALSE, option to overwrite existing articles with the same URL.}
}
\value{
The full path of the updated CSV.
}
\description{
This function creates the CSV if it doesn't exist and appends new articles to it,
ensuring no duplicate URLs are added. If \code{overwrite = TRUE}, it replaces articles
with matching URLs instead of adding duplicates.
}
