% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_schema_elements.R
\name{write_schema}
\alias{write_schema}
\title{Write a new schema to the local CSV file}
\usage{
write_schema(
  website_url,
  author_element,
  title_element,
  date_element,
  text_element
)
}
\arguments{
\item{website_url}{A character string representing the URL of the website.}

\item{author_element}{A character string representing the CSS selector or XPath for the author element.}

\item{title_element}{A character string representing the CSS selector or XPath for the title element.}

\item{date_element}{A character string representing the CSS selector or XPath for the published date element.}

\item{text_element}{A character string representing the CSS selector or XPath for the article text element.}
}
\value{
TRUE if the schema is successfully written, FALSE otherwise.
}
\description{
This function writes a new schema to the local CSV file. If the schema already exists,
it prompts the user for confirmation to overwrite.
}
